Split utils to avoid cross-contaminating the frontend with prisma references when bundling with Vite.
