+++ backend/tests/bookstores.test.js
@@ -7,16 +7,19 @@ import {
   deleteBookstore
 } from "../routes/adminRoutes.js";
 import { prisma } from "../prisma/client.js";
+import { getForMonth } from "../utils.js"
 import {
-  getForMonth,
   createAuthor,
   createBook,
   createBookstore,
   createInventory,
   createPayment,
   createSale,
+  createKindleSale,
+  createCost,
+  createImpression,
   deleteFromDB 
-} from "../utils.js"
+} from "../testUtils.js";
 
 // GETTING
 describe("getting all valid bookstores", async() => {
