+++ backend/tests/inventories.test.js
@@ -11,17 +11,19 @@ import {
   deleteInventory
 } from "../routes/adminRoutes.js";
 import { prisma } from "../prisma/client.js";
+import { getForMonth } from "../utils.js"
 import {
-  getForMonth,
   createAuthor,
   createBook,
   createBookstore,
   createInventory,
   createPayment,
   createSale,
+  createKindleSale,
+  createCost,
   createImpression,
   deleteFromDB 
-} from "../utils.js"
+} from "../testUtils.js";
 
 // GETTING
 describe("getting all valid inventories", () => {
