+++ backend/tests/sales.test.js
@@ -1,17 +1,19 @@
 import { describe, expect, vi, it, beforeAll, afterAll } from "vitest";
 import { getSales, addSale, updateSale, deleteSale } from "../routes/adminRoutes.js";
 import { prisma } from "../prisma/client.js";
+import { getForMonth } from "../utils.js"
 import {
-  getForMonth,
   createAuthor,
   createBook,
   createBookstore,
   createInventory,
   createPayment,
   createSale,
+  createKindleSale,
+  createCost,
   createImpression,
   deleteFromDB 
-} from "../utils.js"
+} from "../testUtils.js";
 
 async function seed() {
   let newAuthor = await createAuthor(prisma, "G", "B", 'g.b@gmail.com', "author");
