+++ frontend/src/BookInventory.jsx
@@ -330,9 +330,9 @@ function BookInventory({
       case 'edit':
         setModalAction("edit");
         break;
-      // case 'delete':
-      //   setModalAction("delete");
-      //   break;
+      case 'delete':
+        setModalAction("delete");
+        break;
       default:
         console.log("Unknown error")
         return;
